+++ SkiaSharpDemo/SkiaSharpDemo/Graphics/Shapes/Line.cs
@@ -8,21 +8,21 @@ namespace SkiaSharpDemo.Graphics
 	{
 		private SKPath path;
 
-		public float X1 { get; set; } = 0.0f;
+		public double X1 { get; set; } = 0.0f;
 
-		public float Y1 { get; set; } = 0.0f;
+		public double Y1 { get; set; } = 0.0f;
 
-		public float X2 { get; set; } = 0.0f;
+		public double X2 { get; set; } = 0.0f;
 
-		public float Y2 { get; set; } = 0.0f;
+		public double Y2 { get; set; } = 0.0f;
 
 		public override SKPath GetPath()
 		{
 			if (path == null)
 			{
 				path = new SKPath();
-				path.MoveTo(X1, Y1);
-				path.LineTo(X2, Y2);
+				path.MoveTo((float)X1, (float)Y1);
+				path.LineTo((float)X2, (float)Y2);
 			}
 
 			return path;
