+++ SkiaSharpDemo/SkiaSharpDemo/Graphics/Shapes/Rectangle.cs
@@ -6,20 +6,20 @@ namespace SkiaSharpDemo.Graphics
 	{
 		private SKPath path;
 
-		public float RadiusY { get; set; } = 0.0f;
+		public double RadiusY { get; set; } = 0.0f;
 
-		public float RadiusX { get; set; } = 0.0f;
+		public double RadiusX { get; set; } = 0.0f;
 
 		public override SKPath GetPath()
 		{
 			if (path == null)
 			{
 				path = new SKPath();
-				var rect = SKRect.Create(0, 0, Width, Height);
+				var rect = SKRect.Create(0, 0, (float)Width, (float)Height);
 
 				if (RadiusX > 0 || RadiusY > 0)
 				{
-					path.AddRoundedRect(rect, RadiusX, RadiusY);
+					path.AddRoundedRect(rect, (float)RadiusX, (float)RadiusY);
 				}
 				else
 				{
